Refactors the CLI so that only main depends on docopt.
